fix no questions yet - closing #16

+++ mini-framework/app/views/user_questions.view.php
@@ -18,9 +18,18 @@
     </div>
     <div id="question-list">
     <?php 
-    foreach ($questions as $question) {
+    if ($questions != null) 
+    {
+      foreach ($questions as $question) 
+      {
         echo $question->asHtml() . "<a href=\"edit?" . htmlentities($question->getId()) . "\">Edit</a>" . "<a href=\"delete?" . htmlentities($question->getId()) . "\">Delete</a>";
+      }
+    }
+    else
+    {
+      echo "<p>No questions yet.</p>";
     }?>
+    </div>
     </div>
   </div>
